Explictly only use Socket::AF_INET
It turns out that the ruby UDPSocket class is perfectly capable of resolving its own domain. The problem is that the Celluloid::DNS::address_family method blew up when we gave it a hostname and not an address.

But, we can just give it Socket::AF_INET all the time and it will work fine. When we are forced to move to IPV6 (god knows when)this might be a problem, but since we're only using this library to talk to route53 this is not an issue.